% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{fizzbuzz}
\alias{fizzbuzz}
\title{FizzBuzz}
\usage{
fizzbuzz(xs, m1 = 3, m2 = 5, silent = opt("quiet"))
}
\arguments{
\item{xs}{a numeric vector of values to consider}

\item{m1}{"buzz" dividend}
}
\description{
You can also cherry-pick options that you want to use, or rename them to
suite your functions.
}
