% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{show_option}
\alias{show_option}
\title{Show an option value as it would be used within this package}
\usage{
show_option(x, ...)
}
\arguments{
\item{x}{An option name to explore}
}
\value{
the option value as it would be discovered inside this package
}
\description{
Explore what value is picked up. You can try setting environment variables or
global options to see how they affect the internally used option value.
}
\examples{
show_option("quiet")
# [1] FALSE

Sys.setenv(R_OPTIONS_EXAMPLE_QUIET = 3)
show_option("quiet")
# [1] "3"

options(options.example.quiet = TRUE)
show_option("quiet")
# [1] TRUE

}
