% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{fizzbuzz}
\alias{fizzbuzz}
\title{FizzBuzz}
\usage{
fizzbuzz(xs, m1 = 3, m2 = 5, silent = opt("quiet"))
}
\arguments{
\item{xs}{a numeric vector of values to consider}

\item{m1}{"buzz" dividend}

\item{silent}{Whether printing should be quiet or loud. See how it's used throughout this
package! (Defaults to \code{NULL}, overwritable using option 'options.example.quiet' or environment variable 'r_options_example_quiet')}
}
\description{
You can also cherry-pick options that you want to use, or rename them to
suite your functions.
}
