% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{options}
\alias{options}
\title{options.example Options}
\usage{
options("options.example.<option>")

options("options.example.<option>" = <value>)
}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Options}{

\describe{
\item{detailed}{\describe{
An example of a more detailed mechanism of defining options, which gives further control over option names, environment variable names and how environment variables are internalized as values.\item{default: }{\preformatted{paste0("tri", "cera", "tops")}}\item{option: }{options.example.detailed}\item{envvar: }{R_OPTIONS_EXAMPLE_DETAILED (as parsed and evaluated expression)}}}
\item{use_options}{\describe{
Define options within your package. Using \code{define} shorthand syntax will auto-generate global R option names and environment variables names to fall back to.\item{default: }{\preformatted{TRUE}}\item{option: }{options.example.use_options}\item{envvar: }{R_OPTIONS_EXAMPLE_USE_OPTIONS (raw)}}}
\item{another_option}{\describe{
Additional options can be chained together. They're also only evaluated on first use, so you're free to use package functions without worrying about order of execution.\item{default: }{\preformatted{exists("a", .GlobalEnv)}}\item{option: }{options.example.another_option}\item{envvar: }{R_OPTIONS_EXAMPLE_ANOTHER_OPTION (raw)}}}
\item{quiet}{\describe{
Whether printing should be quiet or loud\item{default: }{\preformatted{FALSE}}\item{option: }{options.example.quiet}\item{envvar: }{R_OPTIONS_EXAMPLE_QUIET (raw)}}}
}
}

