% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{options_params}
\alias{options_params}
\title{Generated Package Options as Params}
\description{
Here we create a small stub of a function. It's used to inherit parameters
from throughout the package. With this function as a basis, you can use
roxygen2's \verb{@inheritParams} tag to document any parameters that share the
name with an option.
}
